Packaging and refactoring
