+++ config.py
@@ -16,8 +16,9 @@ class TestStatus(Enum):
     
 __version = "1.0.0"
 TEST_PATH = "/data/apps/tests/"
+TEST_CONFIG = "/apps/apptests/tests_config.yaml"
+TRAP_FILE = "/apps/apptests/job_handler/trap.sh"
+TEMP_TEST_FILE = '/data/apps/tests/apptests/temp/'
 REPORT_PATH = "/data/apps/tests/apptests/Reports/"
-STATIC_OUTPUT = ""
-DYNAMIC_OUTPUT = ""
 RUNNING_TESTS = []
 OUTPUT_FIELDS = ["All", "Module", "Dependency", "Time", "JobStatus", "TestStatus", "JobId", "ExitCode", "TetFile"]
