+++ job_handler/__init__.py
@@ -1,0 +1,5 @@
+from test_class import Test
+import config
+
+JobStatus = config.JobStatus
+TestStatus = config.TestStatus
