+++ main.py
@@ -5,38 +5,11 @@
 
 import argparse
 import sys
-import os
-import time
-import threading
 import logging
-import yaml
 from loguru import logger
+
 import config
-import reportHandler.report_generator as report_generator
-import subprocess
-import shlex
-import json
-import hashlib
-
-JobStatus = config.JobStatus
-TestStatus = config.TestStatus
-
-testStatus = {}
-
-class Test(object):
-    def __init__(self, module, dependencies, filepath, startTime, endTime,  jobId, proc, jobStatus, jobOut, jobErr, testStatus, exitCode):
-        self.module = module
-        self.dependencies = dependencies
-        self.filepath = filepath
-        self.startTime = startTime
-        self.endTime = endTime
-        self.jobId = jobId
-        self.proc = proc
-        self.jobStatus = jobStatus
-        self.jobOut = jobOut
-        self.jobErr = jobErr
-        self.testStatus = testStatus
-        self.exitCode = exitCode
+from job_handler import tester
 
 def parse_args(print_help=False):
     class MyParser(argparse.ArgumentParser):
@@ -130,323 +103,6 @@ def _setup_logging(debug, verbose):
         logger.add(sys.stderr, level=level)
     logger.debug("Logging level set to : {}", level)
 
-def checkTestStatus(jobId):
-    '''
-        Returns the status and the exit code of the job with jobId
-    '''
-    global testStatus
-    cmd = ["sacct", "-j", jobId, "-n", "-p", "--format=State,ExitCode"]
-    with subprocess.Popen(
-                cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True
-    ) as proc:
-        res_stdout = proc.stdout.read()
-        res_stderr = proc.stderr.read()
-        testStatus[jobId] = [proc, res_stdout, res_stderr]
-
-    subp = testStatus[jobId]
-    while(subp[0].poll() is None):
-        return ()
-    status = subp[1].split('|')[0]
-    if(subp[2] or status == "RUNNING"):
-        return ()
-    else:
-        exitCode = ""
-        try :
-            exitCode = subp[1].split('|')[1]
-        except:
-            exitCode = ""
-
-        return (status, exitCode)
-
-def get_raw_json(args):
-    '''
-        Adopted from code written by Alex in /apps/lmod/admin/
-        Returns a dict with lmod data
-    '''
-    spider = "/apps/lmod/lmod/libexec/spider"
-    modulepath_root = os.environ["MODULEPATH"]
-
-    print("Info: Reading latest lmod data from "+ modulepath_root)
-    """Retrieve raw json data from either spider output or from a test file."""
-
-    cmd = "{} -o spider-json {}".format(spider, modulepath_root)
-    if args.debug:
-        logger.debug("Cmd: '{}'".format(cmd))
-    cmd_args = shlex.split(cmd)
-    with subprocess.Popen(
-        cmd_args,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        universal_newlines=True,
-        encoding="utf-8",
-    ) as proc:
-        res_stdout = proc.stdout.read().strip()
-        res_stderr = proc.stderr.read()
-    if res_stderr:
-        print(
-            "Error: the getent command call was not successful. Submit a support request"
-        )
-        if args.debug:
-            logger.debug(res_stderr)
-    if args.debug and args.verbose:
-        logger.debug("stdout: '{}'".format(res_stdout))
-    if not res_stdout:
-        logger.error("Error: No output from the spider command.")
-    raw_data = json.loads(res_stdout)
-    if args.debug:
-        logger.debug(raw_data)
-    return raw_data
-
-def checkJobStatus(args):
-    '''
-        Checks test status and updates the report
-    '''
-    global JobStatus
-    global TestStatus
-
-    # A thread that starts all test jobs and puts them in the config.RUNNING_TESTS list
-    thread = threading.Thread(target=startTests(args))
-    thread.start()
-    thread.join()
-    logger.debug("startTests Done")
-
-    # A dict that keeps track of the completed tests so far
-    testsCompleted = {}
-
-    # Report is generated with the current timestamp
-    report_name = time.strftime("%Y-%m-%d-%H_%M_%S")
-
-    # Loops until all test jobs submitted complete with some status
-    while(len(testsCompleted) <= len(config.RUNNING_TESTS)):
-        logger.debug("The number of completed tests are: " + str(len(testsCompleted)) + "/" + str(len(config.RUNNING_TESTS)))
-        logger.debug("The number of tests running are: " + str(len(config.RUNNING_TESTS) - len(testsCompleted)))
-
-        # Processes the status of each job submitted and puts them appropriately in testsCompleted
-        for job in config.RUNNING_TESTS:
-            #  If module name is invalid or test file is missing, add this job to 'testsCompleted' with respective status
-            if(job.module not in testsCompleted and (job.jobStatus == JobStatus.MISSING or job.jobStatus == JobStatus.INVALID)):
-                logger.debug(job.jobStatus)
-                # if(os.path.exists(job.filepath)):
-                #     removeTrap(job.filepath)
-                testsCompleted[job.module + str(job.dependencies)] = ()
-                continue
-            # If job is submitted, check if there is any new status
-            elif(job.jobStatus is not JobStatus.MISSING and job.proc.poll() is not None):
-                if(job.proc.returncode == 0):
-                    job.jobStatus = JobStatus.SUBMITTED
-                    job.testStatus = TestStatus.RUNNING
-                else :
-                    job.jobStatus = JobStatus.FAILED
-                    job.testStatus = TestStatus.NA
-
-                # Process the job status
-                jobId = job.jobOut.split()[3]
-                job.jobId = jobId
-                testStatus = checkTestStatus(jobId)
-                if(len(testStatus) > 0):
-                    job.exitCode = testStatus[1]
-                    if(testStatus[0] == TestStatus.COMPLETED.value):
-                        job.testStatus = TestStatus.COMPLETED
-                    elif(testStatus[0] == TestStatus.FAILED.value):
-                        job.testStatus = TestStatus.FAILED    
-
-                    if(job.testStatus is not TestStatus.RUNNING and job.module not in testsCompleted):
-                        # removeTrap(job.filepath)
-                        testsCompleted[job.module + str(job.dependencies)] = testStatus
-
-        # After processing the entire batch of test jobs for new upates, genrate the latest report
-        report_generator.generateReport(report_name, args)
-        if(len(testsCompleted) == len(config.RUNNING_TESTS)):
-            report_generator.generateReport(report_name, args, exit=True)
-            print("Success: Program completed")
-            print("Success: Check the report - /data/apps/tests/apptests/Reports/" + report_name + ".txt (.json also available)")
-            print("Success: Check slurm logs for a particular test - /data/apps/tests/apptests/slurm_logs/<job_id>.log\n")
-            logger.debug("Waiting for thread2")
-            sys.exit(0)
-
-def read_bash_code(file_path):
-    with open(file_path, 'r') as file:
-        bash_code = file.read()
-    return bash_code
-
-def addTrap(file_path):
-    bash_code = read_bash_code(os.path.join(os.getcwd(), 'trap.sh'))
-    with open(file_path, 'r') as file:
-        lines = file.readlines()
-    modified_lines = []
-    addedTrap = False
-
-    for line in lines:
-        if line.startswith('# BEGIN'):
-            return
-        if addedTrap or line.startswith('#') or line.startswith('\n'):
-            modified_lines.append(line)
-        else:
-            modified_lines.append(bash_code + "\n")
-            modified_lines.append(line)
-            addedTrap = True
-
-    input_bytes = file_path.encode('utf-8')
-    # Compute the SHA-256 hash
-    sha_hash = hashlib.sha256(input_bytes)
-
-    # Return the hexadecimal representation of the hash
-    newPath = '/data/apps/tests/apptests/temp/' + sha_hash.hexdigest()
-
-    with open(newPath, 'w') as file:
-        file.writelines(modified_lines)
-    return newPath
-
-def removeTrap(file_path):
-    with open(file_path, 'r') as file:
-        lines = file.readlines()
-
-    modified_lines = []
-    inTrapBlock = False
-
-    for line in lines:
-        if line.startswith('# BEGIN'):
-            inTrapBlock = True
-        if line.startswith('# END'):
-            inTrapBlock = False
-            continue
-
-        if not inTrapBlock:
-            modified_lines.append(line)
-
-    with open(file_path, 'w') as file:
-        file.writelines(modified_lines)
-
-def submitAllJobs(lmod, yaml_config):
-    for module in lmod:
-    #    submitJob(lmod, yaml_config, module)
-        pass
-
-
-def submitJob(lmod, yaml_config, module, _moduleVersion = None):
-    '''
-        Submits slurm jobs for valid modules provided
-    '''
-    logger.debug(_moduleVersion)
-
-    # Check if valid module name
-    if module not in lmod.keys() :
-        print("Error: Invalid module '" + module + "'")
-        print("Warning: Maybe you have entered '<module>/<version>' with a '-m' flag. Use '-mv' instead.")
-
-        config.RUNNING_TESTS.append(Test(module, "-", "-", time.time(), time.time(), 'N/A', None, JobStatus.INVALID, None, None, TestStatus.MISSING, "-"))
-        return
-
-    # Check if valid module/version
-    if _moduleVersion is not None :
-        moduleVersion = [(module, luaPath) for module in lmod for luaPath in lmod[module] if _moduleVersion == lmod[module][luaPath]['fullName']]
-        logger.debug(moduleVersion)
-        if len(moduleVersion) == 0 :
-            print("Error: Invalid version '" + _moduleVersion + "'")
-            print("Warning: Maybe you have entered '<module>' with a '-mv' flag. Use '-m' instead.")
-
-            config.RUNNING_TESTS.append(Test(_moduleVersion, "-", "-", time.time(), time.time(), 'N/A', None, JobStatus.INVALID, None, None, TestStatus.MISSING, "-"))
-            return
-
-    logger.debug("Valid module or module/version " + str(_moduleVersion))
-
-    # Parse lmod dict for dependency details
-    for luaPath in lmod[module]:
-        moduleVersion = lmod[module][luaPath]["fullName"]
-        logger.debug(moduleVersion)
-        if(_moduleVersion is None or moduleVersion == _moduleVersion):
-            testFilePath = os.path.join(config.TEST_PATH, module, "run.sh")
-
-            # Arguments to the sbatch command
-            args = []
-            dependencies = "-"
-
-            # Check if custom test config (in tests_config.yaml) has been provided and parse it
-            if(moduleVersion in yaml_config.keys()):
-                testFilePath = yaml_config[moduleVersion]["path"]
-                print("Info: Using custom file path " + testFilePath +" for " + moduleVersion)
-                if("args" in yaml_config[moduleVersion]):
-                    for arg in yaml_config[moduleVersion]["args"].split(' '):
-                        args.append(arg)
-                    print("Info: Using custom dependency " + str(args) + " for " + moduleVersion)
-                    dependencies = "Custom config"
-
-            logger.debug(testFilePath)
-
-            # Add all dependencies to args if not in test config
-            if len(args) == 0 and "parentAA" in lmod[module][luaPath]:
-                for dependency in lmod[module][luaPath]["parentAA"][0]:
-                    if(len(dependencies) == 1):
-                        dependencies = dependency
-                    else :
-                        dependencies += "," + dependency
-                    args.append(dependency)
-                    
-            logger.debug(dependencies)
-
-            # Submit test job
-            if(os.path.exists(testFilePath) and os.access(testFilePath, os.R_OK)):
-                # Setting trap conditions and module loads in the test files
-                newPath = addTrap(testFilePath)
-
-                cmd = ['sbatch', newPath]
-
-                args.append(moduleVersion)
-                cmd.extend(args)
-                logger.debug(cmd)
-                with subprocess.Popen(
-                    cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True
-                ) as proc:
-                    res_stdout = proc.stdout.read()
-                    res_stderr = proc.stderr.read()
-                submitTime = time.time()
-                logger.debug(res_stdout)
-                logger.debug(res_stderr)
-                logger.debug("EXIT Code: {}".format(proc.returncode))
-
-                # Add the current job to the RUNNING_TESTS list
-                config.RUNNING_TESTS.append(Test(moduleVersion, dependencies, testFilePath, submitTime, None, 'N/A', proc, JobStatus.PENDING, res_stdout, res_stderr, TestStatus.NA, "-"))
-            else:
-                print("Error: Missing test file /data/apps/tests/" + module + "/run.sh for " + moduleVersion)
-                config.RUNNING_TESTS.append(Test(moduleVersion, dependencies, testFilePath, time.time(), None, 'N/A', None, JobStatus.MISSING, None, None, TestStatus.MISSING, "-"))
-
-def startTests(args): 
-    '''
-        Loads lmod data, custom yaml configs and calls submitJob() for the module(s) to be tested
-    '''
-
-    # Reading module list from spider  
-    lmod = get_raw_json(args)
-
-    '''Alternate way to read lmod data, make sure to refresh the file with latest data first'''
-    #with open("/apps/apptests/lmod_spider_output.txt") as fh:
-    #   lmod = json.load(fh)
-    #  print("Warning: Using lmod data from lmod_spider_output.txt which might be outdated")
-    
-    # Reading test configurations from tests_config.yaml
-    with open("/apps/apptests/tests_config.yaml", "r") as stream:
-        try:
-            config = yaml.safe_load(stream)
-        except yaml.YAMLError as exc:
-            logger.debug(exc)
-            print("YAML config file error. Please check the file.")
-
-    # Tests all versions of the specified modules
-    if(args.module):
-        for module in args.module:    
-            logger.debug(module)      
-            submitJob(lmod, config, module)
-    
-    # Tests a specific module/version 
-    if(args.moduleversion):
-        for arg in args.moduleversion:
-            logger.debug(arg) 
-            submitJob(lmod, config, arg.split('/')[0], arg)
-    
-    if(args.testall):
-        submitAllJobs(lmod, config)
-                 
-
 def main():
     args = parse_args()
 
@@ -466,7 +122,9 @@ def main():
     logger.debug(args.module)
 
     print("Info: Testing is about to begin. This may take a while to finish.")
-    checkJobStatus(args)
+
+    # Start tests
+    tester.startTests(args)
 
 if __name__ == '__main__' :
     main()
