+++ report_handler/__init__.py
@@ -1,0 +1,4 @@
+import config
+
+JobStatus = config.JobStatus
+TestStatus = config.TestStatus
